Clock Speed Change
Changed Sysclk back to 80Mhz and HCLK is now 80MHz as well. Changed PLLSAIR1 which connects to adc mux back to 80MHz and LPUART1 and I2C2 mux, which both use pllclk is 40MHz. SPI baud rate is now 20MBit/s, which is above the requirement of 12-16MHz for Open Alliance, however in ADIN1110 documnetation it says it can be greater than 16MHz